requirements.txt
future==0.15.2
oauthlib==1.1.2
Pillow==3.3.0
pytesseract==0.1.6
python-twitter==3.1
requests==2.10.0
requests-oauthlib==0.6.2
spotipy==2.3.8